chore: move gatsby cloud url detection into own file
